Refactor Userpage
ref: getShoppingpage function and getShopData and Lessons
